+++ framework/jsx-runtime.d.ts
@@ -1,12 +1,33 @@
 export namespace JSX {
-  type Element = unknown
+  type Element = any
   type Children = unknown
+
+  type TargetedEvent<Target extends EventTarget, NativeEvent extends Event = Event> = NativeEvent & {
+    currentTarget: Target
+    target: EventTarget & Target
+  }
+
+  type IntrinsicProps<Target extends EventTarget> = Record<string, unknown> & {
+    children?: Children
+    onBlur?: (event: TargetedEvent<Target, FocusEvent>) => unknown
+    onChange?: (event: TargetedEvent<Target>) => unknown
+    onClick?: (event: TargetedEvent<Target, MouseEvent>) => unknown
+    onFocus?: (event: TargetedEvent<Target, FocusEvent>) => unknown
+    onInput?: (event: TargetedEvent<Target, InputEvent>) => unknown
+    onKeyDown?: (event: TargetedEvent<Target, KeyboardEvent>) => unknown
+    onKeyUp?: (event: TargetedEvent<Target, KeyboardEvent>) => unknown
+    onSubmit?: (event: TargetedEvent<Target, SubmitEvent>) => unknown
+  }
 
   interface IntrinsicAttributes {
     key?: string | number
   }
 
-  interface IntrinsicElements {
+  type IntrinsicElements = {
+    [Name in keyof HTMLElementTagNameMap]: IntrinsicProps<HTMLElementTagNameMap[Name]>
+  } & {
+    [Name in keyof SVGElementTagNameMap]: IntrinsicProps<SVGElementTagNameMap[Name]>
+  } & {
     [elementName: string]: Record<string, unknown>
   }
 }